Adding the main logic to recieve and response the request thats are goin to handle the whatsapp bot
We adding a full context class that pass  a couple of setting that we are goin to send to the ollama api with the AI Model  Llama3.2

The idea of these project is to make a AIBot that can handle responses to questions or messages about Software Devlopment, best practices and the standard of normatives to create scalable, maintenable and efficient software development solutions.

We separate the logic of the Whastapp Services and the Ollama Services for a better code structure

Also add an PromptContextService, this because we need to manage and design the promp that we need to pass to the AI Llama3.2  and delimit his capacity of response in the desired area, so, any kind of mesage that recieve the model can be rejected if the question or message the we are passing by whatsapp doesn't fit with the full context that we are passing in these class.

 Finally in a new controller called RecieveMessageController , the backend recieves an message from whatsapp and we save the message and phone in variables to in the next lines, the message has going to be used for create the prompt that we pass to the previus explained service class,  then,  when we recieve the prompt with a context delimitated we pass these prompt to the OllamaService and get a responseMessage that were going to pass for the whatsapp user that make the request by his phone number